optimize updating of large containers
The container view currently manages its contents via a vector.

While this is simple and sufficient for a small to medium number of
instances, it results in pool scaling for updates of larger containers.

To resolve this, an unordered map can be used to for fast lookup during
updates, while a deque (holding only references) is used for sorting
and subsequent rendering.

Note that a deque was chosen over a vector, because profiling has shown
significant benefits when instance are removed.