properly show nested instances in container view
The container view currently only displays properties of an instance
itself.

This makes nested instances (i.e. those with properties of struct
types) to be difficult to read, because they are just displayed as a
string. This also means that it is currently not possible to sort
instances by a sub-struct property.

Instead, nested instances can be 'flattened' and sub-properties
displayed as columns of the top-level instance.

Note that this can currently only be done up to a limit of 64 total
properties, due to a hard column limit by Dear ImGui.