prevent potential UB when settings are destroyed
When the app is closed, the static settings storage, containing DOTS
instances, can currently be destroyed after destruction of the static
descriptor map, which results in undefined behavior, due to missing
type descriptors.

To avoid this, the settings can explicitly be cleared before static
cleanup is started.

Note that the backend has to be destroyed explicitly before clearing
the settings to ensure they are persisted before being removed.