extract DOTS data models
Most widgets currently directly handle the translation between the DOTS
data and GUI data (mostly text and colors).

While this was sufficient for simple GUI elements, the classes are
starting to get too convoluted to handle, especially when different
types of widgets require the same data.

Instead, the translation can be moved into dedicated model classes,
which abstract the data from particular GUI elements.

Note that this is a straightforward migration at this point and does
not included a lot of simplifications that the data models allow.