fix editing of equally named sub-struct properties
When a type contains sub-structs with identically named properties,
they can currently not be edited properly.

This is because due to the same name the corresponding input fields
have the same Dear ImGui label, which has to be unique.

This can be resolved by incorporating the unique memory address of the
property edit into label.