introduce file connections
Since a DOTS guest does not have to care where events are coming from,
reading trace files can be supported by using a corresponding channel.

Subsequent events can then be processed as usual, which allows the
cache and trace views to be used in the same way as when connected to a
"live" host.

Note that even though the file can be specified via a corresponding
endpoint in the same way as any other host, this commit also adds drag
and drop support for files from outside the application to make this
more user friendly.