avoid application freezes on large event queues
When there are a lot of event handlers (i.e. tens of thousands or more)
scheduled for execution in the Asio IO context, it can currently happen
that the application "freezes" or stalls until the queue is empty.

To avoid this, the event loop can be processed in chunks with regular
checks to determine whether processing should be aborted to render the
next frame.

Note that even though the processing might overall be a little slower,
the interface is more responsive for the user in high-load scenarios.