optimize event filter performance in trace view
The trace view currently always resets the filtered item list when a
new event is received.

This is unnecessarily wasteful, since the filter conditions only have
to be applied to the entire list when the filter condition changes.

Instead, the filter condition can be applied only to the new event
itself.