resolve names for early or outdated publishers
The publisher names are currently determined lazily by inspecting the
current content of the DotsClient container when a text is first
accessed.

Since guests might publish something before the DotsClient container is
filled, publisher names might not be resolved, even though the
publisher information is present at a later point.

The same goes for guests that are no longer connected, but which might
have still active instances in the system.

Instead, publisher names have to be tracked directly every time the
DotsClient container changes.

Note that this change also centralizes the publisher names, which
significantly reduces memory usage and allocations.