store property value color only once
The property value color is currently stored in the property model
itself.

This is unnecessarily wasteful, because the color only depends on the
type and not on the property's state.

Since colors are comparatively large (i.e. 4 * sizeof(float) = 16
bytes), memory usage can be reduced significantly by moving the color
into the property descriptor model.