collect and show updated instance in trace view
For cached types, it is often useful to not only trace the published
instance, but also the updated instance in the container, after the
publish was applied.

Note that the added column is hidden by default and that the updated
instance is also shown in the quick view tooltip, including a highlight
of the updated properties.