fix publishing of sub-struct properties
When publishing, the struct edit currently always uses the the property
index (or "tag") of the property itself to create the "included
properties" set.

In the case of a sub-struct property, this often erroneously adds a
wrong property to the set, resulting in incorrect publishes.

Instead, the set has to be build from the first element of the property
path (i.e. the sub-struct property in the top-level struct).