overhaul indicators for publishing structs
When editing a struct for publishing, the property edit currently only
indicates whether a property has a valid input and whether it will be
included in the publish.

Instead, properties can be marked with a compact indicator that shows
whether or not this property will likely be created, updated or removed
(i.e. invalidated).

Additionally, these indicators can be supported by tooltips that
explain in more detail what the operation will most likely do or why a
publish cannot be performed (e.g. due to missing keys).