fixate Dear ImGui to latest master revision
The project currently references the master branch for the Dear ImGui
dependency.

While the trunk appears to be quite stable, the dependency should still
be fixated to a specific version to ensure reliable builds.

Since the latest official release (tag v1.87) is from February, the
revision can be fixed to the most recent master commit.