avoid console when getting release info on Windows
On Windows, creating a child process by default always creates a new
console window.

This currently results in a confusing popup on startup, when curl is
called to fetch the latest release info.

Since this behavior cannot be changed easily without using non-portable
Win32 calls, alternative curl can be called via Boost Process, which
provides a corresponding option.

Note that this unfortunately incurs another dependency, as Boost
Process relies on Boost Filesystem, which is not header-only.