replace std::regex with re2 library
For a variety of reasons, std::regex is notoriously slow in most
implementations.

While this is often not that significant for simple evaluations, it
makes a dramatic difference for things like the trace view, where
regular expressions have to be applied to potentially tens or hundreds
of thousands of events.

To ensure that the filtering is responsive (or "snappy"), the
implementation must be replaced with something more performant.

While there are a couple of attractive options available (particularly
hyperscan), many of them are usually non-trivial to integrate in a
portable way, due to various dependencies.

Google's re2 library on the other hand, is a comparatively simple,
proven and popular engine, which can easily be obtained and used via
FetchContent_Declare. At the same time, it is still among the fastest
solutions around (see [1]).

Note that due to its design, re2 does not support lookarounds, which,
among other things, makes it impossible to specify inverted filters.

References:

 - [1] https://github.com/HFTrader/regex-performance