introduce filter targets for trace view
Filters are currently exclusively applied to type names.

While this is sufficient for many scenarios, it is sometimes useful to
be able to search for other kinds of information, such as publishers or
even property data.

This commit introduces such functionality to the trace view along with
corresponding options.

Note that currently all filter targets are enabled by default, because
the performance seems to be sufficient even for large traces.