introduce "matcher" abstraction for filtering
Filtering is currently always done via regular expressions.

While regular expressions are often sufficient and flexible, they are
also usually slower than simple sub-string matching.

To allow additional kinds of matching in the future, filtering can be
implemented by introducing a Matcher class, which can be extended to
other matching mechanisms in addition to regex inputs.

Note that this requires an overhaul of the way filter settings are
stored, which is a breaking change for currently stored filter
settings.