introduce 'quick filters'
While the re2 regular expression engine is pretty fast, even for large
traces with tens of thousands of events, it still has a considerable
overhead.

Additionally, negations are not possible with re2, which currently
prevents users from excluding items (e.g. types) from a view.

Since the most common use case for filtering is the simple disjunction
of multiple, possibly negated, sub-strings, it makes sense to support
this directly.

This commit introduces 'quick filters', which can be specified via a
very simple syntax of only using the '|' symbol for disjunctions and
'!' for negations.

They can be implemented by simple string searching and are
significantly faster than regular expressions, particularly when
multiple terms are combined.

Note that regular expressions can still be used, but are no longer the
default and instead have to be enabled manually.