ensure that settings properties are always valid
Since properties of settings objects may not be valid when loaded from
the config file, they have to be initialized manually.

While this was trivial in the past, it starts to get more complicated,
due to the amount of settings and different locations where they are
accessed.

This can be simplified by default constructing all properties when a
settings object is loaded from the file.

Note that this requires some additional logic in some cases where the
desired default setting is different from the default init value.