exclude quotes when editing string properties
The property edit is currently showing the value of string properties
in the same way they are rendered everywhere else, including the
enclosing quotation marks.

This can be confusing to users, because it is not clear whether or not
the quotes have to be specified or not, which may result in
unintentionally publishing a string values containing the quotes.