fix startup error when no config file is present
The application currently crashes on startup when no config file is
present (e.g. when the app is started for the first time).

This issues is caused by improper handling of incomplete config data
and is a regression that was introduced by a commit that was supposed
to prevent this very problem (see [1]).

The current solution is insufficient, because the config object might
be accessed in the first frame, before the settings are loaded from the
config file.

Instead, the object has to be defaulted at the point of registration

References:

 - [1] commit b12a196edc17ae4c8c0da4feed673e4d86fa82d7