fix selection of similar cached instances
In some cases, instances in the cache view can currently not be
selected when there are other instances with similar contents.

This is because ImGui::Selectable uses the label itself as the ID of
the widget when none is specified explicitly, which can result in
entries of multiple rows sharing the same identifier.

Instead, the pointer of the item can be used as the ID explicitly.

Note that since the entire row is supposed to be selected when
selecting a single entry, it is possible to use the same ID for all of
them, as long as it is not shared among different rows.