fix automatic deselection of updated cache items
Items currently lose their selection flag when the instance is updated.

This is a regression that was introduced when the handling and storage
of event data was overhauled (see [1]).

Since the item is currently replaced by a new one when the instance is
updated, the selection information is lost because the flag is stored
in the item object itself.

Instead, struct items need to offer the possibility of updating the
model without having to constructing a new item altogether.

References:

 - [1] commit 5ec489139534eab21f974d99f459847dc1a3eac5