fix instance id collisions
The instance identifier, which is supposed to uniquely identify the
(cached) instance an event updates, is currently generated from the
address of the container instance.

While this makes the id unique for any individual instance until it is
removed, the same id might be used for another instance afterwards.

This causes issues, when removes are not processed immediately, but
instead are deferred for batch processing at a later point in time (as
it is the case for the cache view).

To fix this, the concept can be replaced by a proper 'update index',
which refers to the actual event that created the instance.

Note that in addition to fixing the problem, this approach also allows
the grouping of related events in the future.