fix cache/container size update on remove events
The cache/container size is currently not updated properly, when the
corresponding struct view is closed.

The issue is caused by using the size of the item list instead of the
container, which is only cleaned up when the view is actually open and
being rendered.

Note that this is a regression that was introduced when the handling
and storage of event data was overhauled (see [1]).

References:

 - [1] commit 5ec489139534eab21f974d99f459847dc1a3eac5