use single-line if-else statements where sensible
While single-line if-else statements (i.e. those without a body block)
can be much less verbose, they should usually be avoided to prevent the
pitfall of unintentionally adding statements outside the corresponding
branches.

However, since this kind of "misleading indentation" is warned about by
the GCC compiler when using -Wall (see [1]), as well as detected by
Clang diagnostic analysis (both of which are used when developing this
project), single-line statements can be used in a safe way.

Note that single-line if-else statements must not be nested, due to the
'dangling else' problem (see [2]), which fortunately is also warned
about by GCC.

References:

 - [1] https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html

 - [2] https://en.wikipedia.org/wiki/Dangling_else