simplify widget id handling for item context menus
To differentiate the context menus of individual struct or trace items
respectively, ImGui::BeginPopupContextItem requires a unique widget
identifier that has to be passed as a string.

These strings are currently generated and stored in each item, which is
wasteful, particularly when opening large trace files.

While ImGui::BeginPopupContextItem only accepts strings and does not
work with ImGui::PushID, the implementation is quite small and can be
provided as an extension in a modified version that accepts pointers
instead.