fix editing and removal of stored filters
Filter settings for both the cache and trace view can currently neither
be edited nor removed.

This is caused by an inversion of the check for the corresponding menu
entries, which results in the options not being present when they
should be and in undefined behavior (most often a crash) when they are
used.

Note that this is a regression that was introduced when the filter
validity was addressed (see [1]).

References:

 - [1] commit b12a196edc17ae4c8c0da4feed673e4d86fa82d7