silence console output in release mode
The application currently prints various kinds of non-critical
information when it is started from the console on non-Windows
platforms in release mode.

Since this is a GUI application, which should be able to handle
expected errors, the output should be reduced to an absolute minimum of
critical issues.

Note that this includes silencing the (error) output of the Curl call
for obtaining the version information.