support connections without storing host info
Establishing a connection currently always requires creating a
persistent host setting with an endpoint and description.

This can quickly clutter the host list with one-off connections;
particularly when dealing with trace files.

Instead, users should be able to just enter endpoints manually and only
store them when necessary.

Note that this implementation is based on the filter input, which
already works in a similar manner.