start in last directory when opening file dialog
The file dialog currently always starts in the working directory of the
process.

While this is a decent default, it is likely that users want to use
specific locations repeatedly; particularly for trace files.

The file dialog should therefore persist the last visited location and
use it as a new default accordingly.

Note that this is a non-critical user setting, that may be broken by
future changes.