fix startup crash due to missing timezone database
When the IANA timezone database (aka tzdata) is missing on a system,
the application currently crashes on startup.

This is a regression that was introduced when the timestamp conversion
performance optimization was added (see [1]).

The issue occurs, because the timezone override call will result in an
exception when no tzdata is available.

This needs to be handled accordingly.

References:

 - [1] commit 5560df3c2e27c3b82e9d3acce85b434e66d7db48