fix loading of dots-record files
dots-record published an DotsRecordHeader, it is defined in dots-cli tools
as internal (which may be wrong). Due to the 'internal'-flag, no descriptor
was written.
This commit adds DotsRecordHeader to the file-registry.
Another missing type is DotsCacheInfo, which is also marked internal, but
it is not used in dots-explorer and therefore not automatically added to
the registry.